Fix index handling when updating rows. Fix queries that use multiple indexes.
Allow marking indexes as accessed through Model instance.

Use an `index` property in Field objects to determine whether or not to index on columns.
This removes the need for the table implementation to know about the ForeignKey class and allows for better extensibility for potential other uses of indexes.

Add tests that ensure indexes are updated correctly.

Improve comments in memoize function.