Use Symbol.for to create symbols in the global symbol registry
Prior to this change, when using redux-orm across JavaScript contexts (such as iframes and secondary windows), each context will have a unique reference for of the symbol for a key. This was leading to failed symbol checks.

One specific use case this change supports is as follows -
1. Use redux-orm in a browser window to create ORM state.
2. Create a selector using an ORM instance as a parameter In a secondary window with its own copy of redux-orm.
3. Run that selector, passing in the state reference from the primary window.

Without Symbol.for, the `isOrmState` check in memoize.js fails, resulting in the selector's result function receiving the ORM state instead of the session instance.